telegram: refactor and backend documentation
Refactor of configuration to allow connection to Telegram test data centers. Refactor of features for the filesystem, implementation of some fields around the filesystem. Fix an issue on the object listing by remote path of the object on read from terminal. Updated options to add properties missing for the objects implementation. Added hashing algorithm for verification of the files on local as the Telegram API supports per chunked SHA-256 hashing. Updated documentation for some methods, others are missing.