telegram: method cache and migration to telegram client type
The client now implements mutex locks on directories and files. Channels and topics are cached for multiple reads, along with the use of fs.pacer to avoid rate limiting from the Telegram API. Context is now shared between methods for easier routine cancellation. The required options for the backend have been updated to work with the new implementation.